+++ src/ctrl/mtp_client.py
@@ -19,7 +19,9 @@ from ctrl.util.CIR import MTPProbeCIR
 from ctrl.util.format import MTPDataFormat
 from ctrl.lib.mtpcommand import MTPcommand
 from ctrl.test.manualProbeQuery import MTPQuery
-from EOLpython.Qlogger.messageHandler import QLogger as logger
+from EOLpython.Qlogger.messageHandler import QLogger
+
+logger = QLogger("EOLlogger")
 
 
 class MTPClient():
@@ -64,8 +66,7 @@ class MTPClient():
         try:
             self.rawfile.close()
         except Exception as err:
-            logger.printmsg("ERROR", err + " Unable to close file " +
-                            self.rawfilename)
+            logger.error(err + " Unable to close file " + self.rawfilename)
 
     def printMenu(self):
         """ List user options """
@@ -118,8 +119,7 @@ class MTPClient():
 
         elif cmdInput == '6':
 
-            logger.printmsg("info", "sit tight - Bline scan typically takes " +
-                            "6 seconds")
+            logger.info("sit tight - Bline scan typically takes 6 seconds")
 
             # Make sure the buffer is clear before starting the scan.
             self.clearBuffer()
@@ -153,7 +153,7 @@ class MTPClient():
             exit(1)
 
         else:
-            logger.printmsg("info", "Unknown command. Please try again.")
+            logger.info("Unknown command. Please try again.")
 
     def cycle(self):
         self.cycleMode = True
@@ -161,7 +161,7 @@ class MTPClient():
         # Initialize probe. Return true if success
         success = self.initProbe()
         if not success:  # Keep trying
-            logger.printmsg("info", "Init failed. Trying again")
+            logger.info("Init failed. Trying again")
             time.sleep(1)  # Emulate manual response time. Prob not needed
             # Move home, then init again, because this is what I do
             self.move.moveHome()
@@ -170,7 +170,7 @@ class MTPClient():
         # Move home. Returns true if successful
         success = self.move.moveHome()
         if not success:  # Keep trying
-            logger.printmsg("info", "Move home failed. Trying again")
+            logger.info("Move home failed. Trying again")
             time.sleep(1)  # Emulate manual response time. Prob not needed
             success = self.move.moveHome()
 
@@ -206,33 +206,29 @@ class MTPClient():
                 exit(1)
 
     def createRawRec(self):
-        logger.printmsg("info", "sit tight - complete scans " +
-                        "typically take 17s")
+        logger.info("sit tight - complete scans typically take 17s")
         firstTime = datetime.datetime.now(datetime.timezone.utc)
 
         # Create a raw record
         raw = self.fmt.createRawRecord(self.move)
-        logger.printmsg("info", "RAW\n" + raw)
+        logger.info("RAW\n" + raw)
 
         # Command finished
         nowTime = datetime.datetime.now(datetime.timezone.utc)
-        logger.printmsg("info", "record creation took " +
-                        str(nowTime-firstTime))
+        logger.info("record creation took " + str(nowTime-firstTime))
 
         # Write raw record to output file
         self.writeRaw(raw + "\n")
 
         writeTime = datetime.datetime.now(datetime.timezone.utc)
-        logger.printmsg("info", "record write took " +
-                        str(writeTime-nowTime))
+        logger.info("record write took " + str(writeTime-nowTime))
 
         # Create the UDP packet
         udpLine = self.fmt.createUDPpacket()
         self.sendUDP(udpLine)
 
         udpTime = datetime.datetime.now(datetime.timezone.utc)
-        logger.printmsg("info", "udp creation took " +
-                        str(udpTime-writeTime))
+        logger.info("udp creation took " + str(udpTime-writeTime))
 
     def writeFileTime(self, time):
         """
@@ -267,12 +263,11 @@ class MTPClient():
             # Move to first angle in readBline
             cmd, currentClkStep = self.fmt.getAngle(80, 0)
             s = self.move.moveTo(cmd, self.data)
-            logger.printmsg('info', "First angle reached = " + str(s))
+            logger.info("First angle reached = " + str(s))
 
             # Command finished
             nowTime = datetime.datetime.now(datetime.timezone.utc)
-            logger.printmsg("info", "single move took " +
-                            str(nowTime-firstTime))
+            logger.info("single move took " + str(nowTime-firstTime))
 
     def readFreqTest(self):
         """ Test (and time) read freq for all channels - no move """
@@ -285,9 +280,8 @@ class MTPClient():
         # Command finished
         nowTime = datetime.datetime.now(datetime.timezone.utc)
 
-        logger.printmsg("info", "data from one position:" + str(countStr))
-        logger.printmsg("info", "freq triplet creation took " +
-                        str(nowTime-firstTime))
+        logger.info("data from one position:" + str(countStr))
+        logger.info("freq triplet creation took " + str(nowTime-firstTime))
 
     def createElineTest(self):
         """ Create E line """