+++ src/proc/plotTimeSeries.py
@@ -11,7 +11,9 @@ import matplotlib.dates as mdates
 from matplotlib.backends.backend_qt5agg import (
        FigureCanvasQTAgg as FigureCanvas)
 from PyQt5.QtWidgets import QComboBox
-from EOLpython.Qlogger.messageHandler import QLogger as logger
+from EOLpython.Qlogger.messageHandler import QLogger
+
+logger = QLogger("EOLlogger")
 
 
 class TimeSeries():
@@ -60,8 +62,7 @@ class TimeSeries():
         # Test and make sure there is data in the flightData array. If empty,
         # prompt user to load some data and return.
         if self.client.reader.getNumRecs() == 0:
-            logger.printmsg("ERROR", "No data available: ",
-                            "Try loading some raw data")
+            logger.error("No data available: ", "Try loading some raw data")
             return()
 
         # Find the date in the data (YYYYMMDD) and convert to base